refactor(api): move handlers to root of routes dir
