+++ api/server/routes/ask/google.js
@@ -4,7 +4,7 @@ const crypto = require('crypto');
 const { titleConvo, GoogleClient } = require('../../../app');
 // const GoogleClient = require('../../../app/google/GoogleClient');
 const { saveMessage, getConvoTitle, saveConvo, getConvo } = require('../../../models');
-const { handleError, sendMessage, createOnProgress } = require('./handlers');
+const { handleError, sendMessage, createOnProgress } = require('../handlers');
 const requireJwtAuth = require('../../../middleware/requireJwtAuth');
 
 router.post('/', requireJwtAuth, async (req, res) => {
