+++ api/server/routes/ask/openAI.js
@@ -4,7 +4,7 @@ const { titleConvo, OpenAIClient } = require('../../../app');
 const { parseConvo } = require('librechat-data-provider');
 const { getAzureCredentials, abortMessage } = require('../../../utils');
 const { saveMessage, getConvoTitle, saveConvo, getConvo } = require('../../../models');
-const { handleError, sendMessage, createOnProgress } = require('./handlers');
+const { handleError, sendMessage, createOnProgress } = require('../handlers');
 const requireJwtAuth = require('../../../middleware/requireJwtAuth');
 
 const abortControllers = new Map();
